requirements.txt
celery==5.0.5
Django==3.0.8
django-debug-toolbar==3.1.1
django-extensions==3.0.9
django-filter==2.3.0
django-model-utils==4.0.0
djangorestframework==3.11.0
drf-rw-serializers==1.0.3
drf-yasg==1.17.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-dotenv==0.14.0
python-binance==0.7.9
redis==3.5.3
telethon==1.17.5
asyncio~=3.4.3
#image-parser:
pytesseract
matplotlib

numpy~=1.19.4
regex~=2020.7.14
asgiref~=3.2.10
Pillow~=7.2.0